feat: Refatorar código para usar routers no FastAPI
