+++ app/main.py
@@ -1,33 +1,6 @@
-import os
-import random
 from fastapi import FastAPI
-from dotenv import load_dotenv
-
-if os.getenv("ENV") is None:
-    load_dotenv()
-
-ENV                     = os.getenv("ENV", "development")
-AUTHOR_SYSTEM           = os.getenv("AUTHOR_SYSTEM", "none")
-VERSION_SYSTEM          = os.getenv("VERSION_SYSTEM", "none")
-DESCRIPTION_SYSTEM      = os.getenv("DESCRIPTION_SYSTEM", "none")
-NAME_SYSTEM             = os.getenv("NAME_SYSTEM", "none")
+from app.routers.monitoring import router as monitoring_router
 
 app = FastAPI()
 
-@app.get("/healthcheck")
-def healthcheck():
-    return { "status": "ok", "env": ENV }
-
-@app.get("/metrics")
-def metricks():
-    return { "cpu_usage": f"{random.randint(10,90)}%", "memory_usage": f"{random.randint(100,800)}MB" }
-
-@app.get("/info")
-def info():
-    return {
-        "name": NAME_SYSTEM,
-        "version": VERSION_SYSTEM,
-        "description": DESCRIPTION_SYSTEM,
-        "author": AUTHOR_SYSTEM,
-        "environment": ENV
-    }
+app.include_router(monitoring_router, prefix="/api")
