+++ app/routers/monitoring.py
@@ -1,0 +1,33 @@
+import os
+import random
+from fastapi import APIRouter
+from dotenv import load_dotenv
+
+if os.getenv("ENV") is None:
+    load_dotenv()
+
+ENV                     = os.getenv("ENV", "development")
+AUTHOR_SYSTEM           = os.getenv("AUTHOR_SYSTEM", "none")
+VERSION_SYSTEM          = os.getenv("VERSION_SYSTEM", "none")
+DESCRIPTION_SYSTEM      = os.getenv("DESCRIPTION_SYSTEM", "none")
+NAME_SYSTEM             = os.getenv("NAME_SYSTEM", "none")
+
+router = APIRouter()
+
+@router.get("/healthcheck")
+def healthcheck():
+    return { "status": "ok", "env": ENV }
+
+@router.get("/metrics")
+def metricks():
+    return { "cpu_usage": f"{random.randint(10,90)}%", "memory_usage": f"{random.randint(100,800)}MB" }
+
+@router.get("/info")
+def info():
+    return {
+        "name": NAME_SYSTEM,
+        "version": VERSION_SYSTEM,
+        "description": DESCRIPTION_SYSTEM,
+        "author": AUTHOR_SYSTEM,
+        "environment": ENV
+    }
