+++ tests/test_healthcheck.py
@@ -8,21 +8,21 @@ sys.path.insert(0, '/app')
 client = TestClient(app)
 
 def test_healthcheck():
-    response = client.get("/healthcheck")
+    response = client.get("/api/healthcheck")
     assert response.status_code == 200
     data = response.json()
     assert data["status"] == "ok"
     assert "env" in data
 
 def test_metrics():
-    response = client.get("/metrics")
+    response = client.get("/api/metrics")
     assert response.status_code == 200
     data = response.json()
     assert "cpu_usage" in data
     assert "memory_usage" in data
 
 def test_info():
-    response = client.get("/info")
+    response = client.get("/api/info")
     assert response.status_code == 200
     data = response.json()
     assert "name" in data
